+++ installnbench2.sh
@@ -127,9 +127,11 @@ echo "nv -d $vhostname -s y -u \"ftpu\$(pwgen -1cnys 31)\""
 nv -d $vhostname -s y -u "ftpu$(pwgen -1cnys 31)"
 s
 
+if [ ! -f /usr/bin/h2load ]; then
 echo "yum -y install nghttp2"
 yum -y install nghttp2
 s
+fi
 
 echo "setup ECDSA SSL self-signed certificate"
 s
@@ -158,6 +160,7 @@ cat /usr/local/nginx/conf/ssl_ecc.conf
 sed -i "s|include \/usr\/local\/nginx\/conf\/ssl_include.conf;|\ninclude \/usr\/local\/nginx\/conf\/ssl_include.conf;\ninclude \/usr\/local\/nginx\/conf\/ssl_ecc.conf;|" /usr/local/nginx/conf/conf.d/${vhostname}.ssl.conf
 ngxreload >/dev/null 2>&1
 
+{
 s
 echo "------------------------------------------------------------------------"
 echo "h2load --version"
@@ -260,6 +263,7 @@ h2load --ciphers=ECDHE-ECDSA-AES256-GCM-SHA384 -H 'Accept-Encoding: br' -c200 -n
 ngxrestart >/dev/null 2>&1
 sleep $SLEEP_TIME
 fi
+} 2>&1 | tee "h2load-nginx-https-${DT}.log"
 
 s
 div
